Extract common test setup to jest config
