+++ packages/react-native-editor/__device-tests__/helpers/utils.js
@@ -1,21 +1,19 @@
 /**
  * External dependencies
  */
-import childProcess from 'child_process';
+const childProcess = require( 'child_process' );
 // eslint-disable-next-line import/no-extraneous-dependencies
-import wd from 'wd';
-import crypto from 'crypto';
-import path from 'path';
-import fs from 'fs';
-
+const wd = require( 'wd' );
+const crypto = require( 'crypto' );
+const path = require( 'path' );
 /**
  * Internal dependencies
  */
-import serverConfigs from './serverConfigs';
-import { iosServer, iosLocal, android } from './caps';
-import AppiumLocal from './appium-local';
+const serverConfigs = require( './serverConfigs' );
+const { iosServer, iosLocal, android } = require( './caps' );
+const AppiumLocal = require( './appium-local' );
 // eslint-disable-next-line import/no-extraneous-dependencies
-import _ from 'underscore';
+const _ = require( 'underscore' );
 
 // Platform setup
 const defaultPlatform = 'android';
@@ -37,8 +35,6 @@ const localIOSAppPath = process.env.IOS_APP_PATH || defaultIOSAppPath;
 
 const localAppiumPort = serverConfigs.local.port; // Port to spawn appium process for local runs
 let appiumProcess;
-let iOSScreenRecordingProcess;
-let androidScreenRecordingProcess;
 
 const backspace = '\u0008';
 
@@ -58,119 +54,6 @@ const isAndroid = () => {
 const isLocalEnvironment = () => {
 	return testEnvironment.toLowerCase() === 'local';
 };
-
-const isMacOSEnvironment = () => {
-	return process.platform === 'darwin';
-};
-
-const IOS_RECORDINGS_DIR = './ios-screen-recordings';
-const ANDROID_RECORDINGS_DIR = './android-screen-recordings';
-
-const getScreenRecordingFileNameBase = ( testPath, id ) => {
-	const suiteName = path.basename( testPath, '.test.js' );
-	return `${ suiteName }.${ id }`;
-};
-
-jasmine.getEnv().addReporter( {
-	specStarted: ( { testPath, id } ) => {
-		if ( ! isMacOSEnvironment() ) {
-			return;
-		}
-
-		const fileName =
-			getScreenRecordingFileNameBase( testPath, id ) + '.mp4';
-
-		if ( isAndroid() ) {
-			if ( ! fs.existsSync( ANDROID_RECORDINGS_DIR ) ) {
-				fs.mkdirSync( ANDROID_RECORDINGS_DIR );
-			}
-
-			androidScreenRecordingProcess = childProcess.spawn( 'adb', [
-				'shell',
-				'screenrecord',
-				'--verbose',
-				'--bit-rate',
-				'1M',
-				'--size',
-				'720x1280',
-				`/sdcard/${ fileName }`,
-			] );
-
-			androidScreenRecordingProcess.stderr.on( 'data', ( data ) => {
-				// eslint-disable-next-line no-console
-				console.log( `Android screen recording error => ${ data }` );
-			} );
-
-			return;
-		}
-
-		if ( ! fs.existsSync( IOS_RECORDINGS_DIR ) ) {
-			fs.mkdirSync( IOS_RECORDINGS_DIR );
-		}
-
-		iOSScreenRecordingProcess = childProcess.spawn(
-			'xcrun',
-			[
-				'simctl',
-				'io',
-				'booted',
-				'recordVideo',
-				'--mask=black',
-				'--force',
-				fileName,
-			],
-			{
-				cwd: IOS_RECORDINGS_DIR,
-			}
-		);
-	},
-	specDone: ( { testPath, id, status } ) => {
-		if ( ! isMacOSEnvironment() ) {
-			return;
-		}
-
-		const fileNameBase = getScreenRecordingFileNameBase( testPath, id );
-
-		if ( isAndroid() ) {
-			androidScreenRecordingProcess.kill( 'SIGINT' );
-			// wait for kill
-			childProcess.execSync( 'sleep 1' );
-
-			try {
-				childProcess.execSync(
-					`adb pull /sdcard/${ fileNameBase }.mp4 ${ ANDROID_RECORDINGS_DIR }`
-				);
-			} catch ( error ) {
-				// Some (old) Android devices don't support screen recording or
-				// sometimes the initial `should be able to see visual editor`
-				// tests are too fast and a recording is not generated. This is
-				// when `adb pull` can't find the recording file. In these cases
-				// we ignore the errors and keep running the tests.
-				// eslint-disable-next-line no-console
-				console.log(
-					`Android screen recording error => ${ error.stdout }`
-				);
-			}
-
-			const oldPath = `${ ANDROID_RECORDINGS_DIR }/${ fileNameBase }.mp4`;
-			const newPath = `${ ANDROID_RECORDINGS_DIR }/${ fileNameBase }.${ status }.mp4`;
-
-			if ( fs.existsSync( oldPath ) ) {
-				fs.renameSync( oldPath, newPath );
-			}
-			return;
-		}
-
-		iOSScreenRecordingProcess.kill( 'SIGINT' );
-
-		const oldPath = `${ IOS_RECORDINGS_DIR }/${ fileNameBase }.mp4`;
-		const newPath = `${ IOS_RECORDINGS_DIR }/${ fileNameBase }.${ status }.mp4`;
-
-		if ( fs.existsSync( oldPath ) ) {
-			fs.renameSync( oldPath, newPath );
-		}
-	},
-} );
 
 // Initialises the driver and desired capabilities for appium
 const setupDriver = async () => {
